Deleted individual parsers and added them to Parsers class Deleted unused number processor Refactored AutoForm and AutoFormInternal
