+++ src/AutoFormInternal.js
@@ -1,8 +1,8 @@
 import React, { Component, PropTypes } from 'react'
 import { reduxForm } from 'redux-form';
-import metadataEvaluator from './metadata/evaluator/metadataEvaluator';
-import modelParser from './metadata/model/modelParser';
-import UIManager from './UIManager';
+import MetadataEvaluator from './metadata/MetadataEvaluator';
+import ModelParser from './metadata/model/ModelParser';
+// import UIManager from './UIManager';
 
 class AutoFormInternal extends Component {
     static propTypes = {
@@ -18,22 +18,25 @@ class AutoFormInternal extends Component {
         fieldLayout: PropTypes.string
     };
 
+    //TODO JS: make this work
     getFactory() {
         const { uiType, componentFactory } = this.props;
 
-        if (!uiType) {
-            return componentFactory;
-        }
-
-        return UIManager.getFactoryPerType(uiType);
+        // if (!uiType) {
+        //     return componentFactory;
+        // }
+        //
+        // return UIManager.getFactoryPerType(uiType);
+        return componentFactory;
     }
 
     buildGroupComponent = () => {
         //Fields: this is not the fields passed from AutoForm. This is generated by ReduxForm. 
         //This object has a property for each field. Each property contains all redux props for the given field
         let { fields, fieldMetadata, layout, values } = this.props;
-        let modelProcessed = modelParser.process(values, fieldMetadata);
-        let fieldMetadataEvaluated = metadataEvaluator.evaluate(fieldMetadata, modelProcessed, '', fields);
+
+        let modelProcessed = ModelParser.process(values, fieldMetadata);
+        let fieldMetadataEvaluated = MetadataEvaluator.evaluate(fieldMetadata, modelProcessed, '', fields);
         let componentFactory = this.getFactory();
 
         return componentFactory.buildGroupComponent({
