+++ src/metadata/model/ModelParser.js
@@ -1,20 +1,31 @@
-import dateTimeParser from './parsers/dateTimeParser';
-import numberParser from './parsers/numberParser';
-import entityParser from './parsers/entityParser';
-import arrayParser from './parsers/arrayParser';
+import Parsers from './Parsers';
 
-class ModelProcessor {
+class ModelParser {
+    static parsers = {};
 
-    constructor() {
-        this.parsers = {};
+    /**
+     * Adds the given filters
+     *
+     * @param parser
+     * @param types
+     */
+    static addProcessor(parser, types) {
+        if (!parser) throw Error('\'filter\' should be truthy');
+        if (!types) throw Error('\'type\' should be truthy');
+        if (!(types instanceof Array)) throw Error('type should be an array');
+
+        types.forEach(type => {
+            this.parsers[type] = parser;
+        });
     }
 
     /**
      * Evaluates the given metadata against the model
+     *
      * @param propertyMetadata - Can be either an object or an array of objects
      * @param model
      */
-    process(model, propertyMetadata) {
+    static process(model, propertyMetadata) {
         if (!model) throw Error('\'model\' should be truthy');
         if (!propertyMetadata) throw Error('Argument \'propertyMetadata\' should be truthy');
         if (propertyMetadata.constructor != Array) throw Error('propertyMetadata should be an array');
@@ -29,36 +40,24 @@ class ModelProcessor {
     }
 
     /**
-     * Adds the given filter
-     */
-    addProcessor(parser, types) {
-        if (!parser) throw Error('\'filter\' should be truthy');
-        if (!types) throw Error('\'type\' should be truthy');
-        if (!(types instanceof Array)) throw Error('type should be an array');
-
-        types.forEach(type => {
-            this.parsers[type] = parser;
-        });
-    }
-
-    /**
      * Filters the given metadata against the model
+     *
      * @param metadata
      * @param model
      */
-    processProperty(metadata, model) {
+    static processProperty(metadata, model) {
         let value = model[metadata.name];
         if (value != null && value != undefined && metadata.type && this.parsers[metadata.type]) { // if there's a parser for the property type
             return this.parsers[metadata.type](metadata, value, this);
         }
+
         return value;
     }
 }
 
-let modelParser = new ModelProcessor();
-modelParser.addProcessor(dateTimeParser, ['datetime', 'date', 'time']);
-modelParser.addProcessor(numberParser, ['int', 'float']);
-modelParser.addProcessor(entityParser, ['entity']);
-modelParser.addProcessor(arrayParser, ['array']);
+ModelParser.addProcessor(Parsers.parseDateTime, ['datetime', 'date', 'time']);
+ModelParser.addProcessor(Parsers.parseNumber, ['int', 'float']);
+ModelParser.addProcessor(Parsers.parseEntity, ['entity']);
+ModelParser.addProcessor(Parsers.parseArray, ['array']);
 
-export default modelParser;
+export default ModelParser;
