--- src/metadata/model/parsers/arrayParser.js
@@ -1,3 +1,0 @@
-export default function parse(metadata, value, modelParser) {
-    return value.map(item => modelParser.process(item, (metadata.fields)));
-}
