--- src/metadata/model/parsers/dateTimeParser.js
@@ -1,13 +1,0 @@
-import { getFormat } from 'redux-autoform-utils/lib/helpers/dateTimeHelpers';
-import { getDateLocalizer } from 'redux-autoform-utils/lib/localization/dateLocalizer';
-
-var localizer;
-
-export default function parse(metadata, value) {
-    if(typeof value == 'string') {
-        if(!localizer) localizer = getDateLocalizer();
-        let format = getFormat(metadata, localizer);
-        return localizer.parse(value, format);
-    }
-    return value;
-}
