--- src/metadata/model/parsers/entityParser.js
@@ -1,3 +1,0 @@
-export default function parse(metadata, value, modelParser) {
-    return modelParser.process(value, metadata.fields, modelParser);
-}
